requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.34.0
sqlalchemy==2.0.36
alembic==1.14.1
psycopg2-binary==2.9.10
anthropic==0.42.0
pyyaml==6.0.2
httpx==0.28.1
python-multipart==0.0.19
python-dotenv==1.0.1
pydantic==2.10.4
pydantic-settings==2.7.1
